Splite the easynet_editor.py into multiple files.
